Docs: Mention returning to killed buffers
This is a really powerful feature that's worth mentioning.  Not only
does it allow returning to file-backed buffers that have been killed,
but it also allows returning to "special" file-backed buffers, like
Info buffers, that get reused to access different parts of files, and,
e.g. help buffers, which get reused to show different help info.